Move ObjectId and friends out to its own module.
